// crates/opsml_server/opsml_ui/src/lib/scripts/types.ts
export interface RegisterUser {
  username: string;
  full_name: string;
  password: string;
  email: string;
  security_question: string;
  security_answer: string;
}
// ...
export interface UserExistsResponse {
  exists: boolean;
  username: string;
}

export interface Readme {
  readme: string;
  exists: boolean;
}

export interface FileSetup {
  fileInfo: Files;
  prevPath: string;
  displayPath: string[];
}

export interface registryPageReturn {
  repos: string[];
  registry: string;
  registryStats: registryStats;
  registryPage: registryPage;
}

export interface securityQuestionResponse {
  question: string;
  exists: boolean;
  error: string;
}